move get_logger to own module
