+++ src/dsp_tools/utils/project_create.py
@@ -15,14 +15,10 @@ from dsp_tools.models.propertyclass import PropertyClass
 from dsp_tools.models.resourceclass import ResourceClass
 from dsp_tools.models.user import User
 from dsp_tools.utils.excel_to_json_lists import expand_lists_from_excel
+from dsp_tools.utils.logging import get_logger
 from dsp_tools.utils.project_create_lists import create_lists_on_server
 from dsp_tools.utils.project_validate import validate_project
-from dsp_tools.utils.shared import (
-    get_logger,
-    login,
-    parse_json_input,
-    try_network_action
-)
+from dsp_tools.utils.shared import login, parse_json_input, try_network_action
 
 logger = get_logger(__name__)
