+++ src/dsp_tools/utils/project_create_lists.py
@@ -5,13 +5,9 @@ from dsp_tools.models.exceptions import BaseError, UserError
 from dsp_tools.models.listnode import ListNode
 from dsp_tools.models.project import Project
 from dsp_tools.utils.excel_to_json_lists import expand_lists_from_excel
+from dsp_tools.utils.logging import get_logger
 from dsp_tools.utils.project_validate import validate_project
-from dsp_tools.utils.shared import (
-    get_logger,
-    login,
-    parse_json_input,
-    try_network_action
-)
+from dsp_tools.utils.shared import login, parse_json_input, try_network_action
 
 logger = get_logger(__name__)
