+++ src/dsp_tools/utils/shared.py
@@ -3,8 +3,6 @@ from __future__ import annotations
 import copy
 import importlib.resources
 import json
-import logging
-import logging.handlers
 import time
 import unicodedata
 from datetime import datetime
@@ -19,41 +17,7 @@ from requests import RequestException
 from dsp_tools.models.connection import Connection
 from dsp_tools.models.exceptions import BaseError, UserError
 from dsp_tools.models.propertyelement import PropertyElement
-
-
-def get_logger(name: str) -> logging.Logger:
-    """
-    Create a logger instance, 
-    set its level to INFO,
-    and configure it to write to a file in the user's home directory.
-    
-    Args:
-        name: name of the logger
-
-    Returns:
-        the logger instance
-    """
-    _logger = logging.getLogger(name)
-    _logger.setLevel(logging.INFO)
-    formatter = logging.Formatter(
-        fmt="{asctime} {filename: <20} {levelname: <8} {message}",
-        style="{"
-    )
-    # a RotatingFileHandler fills "filename" until it is "maxBytes" big, 
-    # then appends ".1" to it and starts with a new file "filename",
-    # fills it until it is "maxBytes" big,
-    # then appends ".1" to it (replacing the old ".1" file)
-    logfile_directory = Path.home() / Path(".dsp-tools")
-    logfile_directory.mkdir(exist_ok=True)
-    handler = logging.handlers.RotatingFileHandler(
-        filename=logfile_directory / "logging.log",
-        mode="a",
-        maxBytes=3*1024*1024,
-        backupCount=1
-    )
-    handler.setFormatter(formatter)
-    _logger.addHandler(handler)
-    return _logger
+from dsp_tools.utils.logging import get_logger
 
 logger = get_logger(__name__)
